extract nextgen
