+++ packages/cli/src/nextgen/ops/index.ts
@@ -1,7 +1,23 @@
-import resolve from './ops'
-import { RunnerConfig, RenderedAction, ActionResult } from './types'
+import type { RenderAttributes, ActionResult, RenderedAction, RunnerConfig } from '../types'
 
-const execute = async (
+export const resolve = async (attributes: RenderAttributes) => {
+  const ops = []
+  if (attributes.to && !attributes.inject) {
+    const add = (await import('./add')).default
+    ops.push(add)
+  }
+  if (attributes.to && attributes.inject) {
+    const inject = (await import('./inject')).default
+    ops.push(inject)
+  }
+  if (attributes.sh) {
+    const shell = (await import('./shell')).default
+    ops.push(shell)
+  }
+  return ops
+}
+
+export const execute = async (
   renderedActions: RenderedAction[],
   args: any,
   config: RunnerConfig
@@ -27,5 +43,3 @@ const execute = async (
 
   return results
 }
-
-export default execute
