+++ packages/cli/src/nextgen/render.ts
@@ -1,33 +1,32 @@
-import { RenderedAction, RunnerConfig } from './types'
 import path from 'path'
 import walk from 'ignore-walk'
-import { RenderResult } from './types'
+
+import type { RenderedAction, RunnerConfig, RenderResult } from './types'
 
 const ignores = [
-  'prompt.js',
-  'index.js',
+  'prompt.ts',
+  'index.ts',
   '.hygenignore',
   '.DS_Store',
   '.Spotlight-V100',
   '.Trashes',
-  'ehthumbs.db',
+  'thumbs.db',
   'Thumbs.db'
 ]
 
-async function getFiles (dir: string): Promise<string[]> {
+function getFiles (dir: string): string[] {
   const files = walk
-    .sync({ path: dir, ignoreFiles: ['.hygenignore'] })
+    .sync({ path: dir, ignoreFiles: ignores })
     .map(f => path.join(dir, f))
-  return await Promise.all(files);
+  return files;
 }
 
 const render = async (
   args: any,
   config: RunnerConfig
 ): Promise<RenderedAction[]> => {
-  let files = await getFiles(args.actionfolder);
+  let files = getFiles(args.actionfolder);
   files = files.sort((a, b) => a.localeCompare(b));
-  files = files.filter(f => !ignores.find(ig => f.endsWith(ig)));
   files = files.filter(file =>
     args.subaction
       ? file.replace(args.actionfolder, '').match(args.subaction)
